Deep_Research/requirements.txt
# Core dependencies
gradio>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-throttle>=1.0.0

# OpenAI integration
openai>=1.0.0

# Web search and scraping
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
markdown>=3.4.0

# Optional: For SerpAPI (if you have API key)
# google-search-results>=2.4.2

# SendGrid for email
sendgrid>=6.10.0